Properly handle missing nodes/ways when parsing PBF files
It is perfectly normal for a way/relation to refer to nodes/ways that are absent
from the file we are currently parsing. The OSM parser skips such nodes/ways
and moves on, while the PBF one crashes.

Make the PBF parser ignore these problematic elements, too.

This hopefully fixes #2.