Switch to a greedy algorithm for multipolygon ring detection
This should fix issue #13 (at least for Croatia). The current algorithm with backtracking is quadratic when the multipolygon is missing some of its ways (e.g., when it is exported as part of a city/country). When this happens for multipolygons with ~100K nodes (e.g., the Adriatic Sea), the importer can get stuck for hours.

I don't remember why I originally implemented backtracking, but for valid multipolygons, just greedily expanding rings segment-by-segment is fine. From my limited testing, I can't actually find any multipolygons that would only work with the old algorithm but not with the new one.